// src/scripts/utils/enums.ts
enum SHAPE_SIZES {
    MIN = 32,
    MAX = 124
}

enum SHAPE_TYPES {
    CIRCLE = 'Circle',
    SQUARE = 'Square',
    TRIANGLE = 'Triangle',
    PENTAGON = 'Pentagon',
    HEXAGON = 'Hexagon',
    ELLIPSE = 'Ellipse',
    BLOB = 'GraphicsBlob',
}

enum SCREEN_SIZES {
    WIDTH = 1000,
    HEIGHT = 600
}

enum STATIC_COLORS {
    BLACK = 0x000000,
    WHITE = 0xFFFFFF,
    YELLOW = 0xFFFF00,
    BLUE = 0x0033FF,
    RED = 0xFF0000,
}

export {SHAPE_SIZES, SHAPE_TYPES, SCREEN_SIZES, STATIC_COLORS}